move record to correct source folder
